requirements.txt
uproot
awkward>=0.12.0
xxhash
lz4
requests>=2.22.0
pyarrow
kafka
confluent_kafka
pympler
pika
minio
pytest
pytest-mock
flake8==3.5
coverage==4.5.2
codecov==2.0.15
